+++ src/node/index.js
@@ -1,6 +1,7 @@
 // @ts-check
 export { decodeImageBytes, decodeImageFile } from './decode.js';
 export { encodePreviewPng } from './preview.js';
+export { planGeneratedOutputs, HEADER_MANIFEST, PREVIEW_MANIFEST } from './manifest.js';
 export { parseIniConfig, parseImagesConfig, resolveImageConfig, IMAGES_CONFIG_TEMPLATE } from './config.js';
 export { buildGlobMatcher, buildImagesIgnoreMatcher, globToRegExpSource } from './ignore.js';
 export { buildImageProject, collectImageEntries, createImagesConfig, writeImageProject } from './project.js';